Fixed type-casting to unsigned char

+++ raw_gradient.cpp
@@ -25,8 +25,8 @@ int main() {
   // Example: Generate a simple gradient
   for ( unsigned int x = 0; x < SIZE; ++x ) {
     for ( unsigned int y = 0; y < SIZE; ++y ) {
-      framebuffer[x][y].r = (unsigned char) x/2;
-      framebuffer[x][y].g = (unsigned char) y/2;
+      framebuffer[x][y].r = ((unsigned char) ((unsigned int) x/2));
+      framebuffer[x][y].g = ((unsigned char) ((unsigned int) y/2));
       framebuffer[x][y].b = 0;
     }
   }
@@ -40,7 +40,7 @@ int main() {
   std::ofstream raw("result.rgb", std::ios::binary);
   raw.write((char*)framebuffer, sizeof(framebuffer));
   raw.close();
-  
+
   // Helper object to save the framebuffer as TGA image
   TGAImage tga(SIZE, SIZE, 3, framebuffer);
   tga.write_tga_file("result.tga", true);
